TECH-349: Separator type organization
